bosminer: api: Refactor structure of tests
